adapt Timeline to DASH 2nd Edition

+++ libdash/libdash/source/mpd/Timeline.cpp
@@ -23,27 +23,27 @@ Timeline::~Timeline   ()
 {
 }
 
-uint32_t    Timeline::GetStartTime     ()  const
+uint64_t    Timeline::GetStartTime     ()  const
 {
     return this->startTime;
 }
-void        Timeline::SetStartTime     (uint32_t startTime) 
+void        Timeline::SetStartTime     (uint64_t startTime) 
 {
     this->startTime = startTime;
 }
-uint32_t    Timeline::GetDuration      ()  const
+uint64_t    Timeline::GetDuration      ()  const
 {
     return this->duration;
 }
-void        Timeline::SetDuration      (uint32_t duration) 
+void        Timeline::SetDuration      (uint64_t duration) 
 {
     this->duration = duration;
 }
-uint32_t    Timeline::GetRepeatCount   ()  const
+int32_t    Timeline::GetRepeatCount   ()  const
 {
     return this->repeatCount;
 }
-void        Timeline::SetRepeatCount   (uint32_t repeatCount)
+void        Timeline::SetRepeatCount   (int32_t repeatCount)
 {
     this->repeatCount = repeatCount;
 }
